fix resp

+++ tracker/udp/announce.go
@@ -118,7 +118,7 @@ func (u *UDPTracker) Announce(announce *Announce, remote *net.UDPAddr) {
 		Interval:      shared.AnnounceInterval,
 		Leechers:      incomplete,
 		Seeders:       complete,
-		Peers:         announce.InfoHash.PeerListBytes(int64(announce.NumWant)),
+		Peers:         announce.InfoHash.PeerListBytes(int(announce.NumWant)),
 	}
 	respBytes, err := resp.Marshall()
 	if err != nil {
